控件的动画功能优化：优化动画的回调函数实现方式，简化实现逻辑。

+++ duilib/Control/CheckBox.h
@@ -953,7 +953,7 @@ void CheckBoxTemplate<InheritType>::PaintText(IRender* pRender)
                 DString textColor = GetSelectedStateTextColor(kControlStateHot);
                 if (!textColor.empty()) {
                     drawParam.dwTextColor = this->GetUiColor(textColor);
-                    drawParam.uFade = (uint8_t)this->GetHotAlpha();
+                    drawParam.uFade = this->GetHotAlpha();
                     pRender->DrawString(textValue, drawParam);
                 }
             }
