+++ duilib/Control/RichEdit_SDL.cpp
@@ -1840,7 +1840,7 @@ void RichEdit::Paint(IRender* pRender, const UiRect& rcPaint)
         if (spDrawRichTextCache != nullptr) {
             //通过缓存绘制
             rcDrawText.Offset(0, m_pTextData->GetTextRectOfssetY());
-            pRender->DrawRichTextCacheData(spDrawRichTextCache, rcDrawText, szScrollOffset, m_pTextData->GetTextRowXOffset(), (uint8_t)GetAlpha());
+            pRender->DrawRichTextCacheData(spDrawRichTextCache, rcDrawText, szScrollOffset, m_pTextData->GetTextRowXOffset(), GetAlpha());
         }
         else if (!richTextDataList.empty()) {
             spDrawRichTextCache.reset();
@@ -1853,7 +1853,7 @@ void RichEdit::Paint(IRender* pRender, const UiRect& rcPaint)
                 ASSERT(pRender->IsValidDrawRichTextCache(rcDrawText, richTextDataList, spDrawRichTextCache));
                 //通过缓存绘制
                 rcDrawText.Offset(0, m_pTextData->GetTextRectOfssetY());
-                pRender->DrawRichTextCacheData(spDrawRichTextCache, rcDrawText, szScrollOffset, m_pTextData->GetTextRowXOffset(), (uint8_t)GetAlpha());
+                pRender->DrawRichTextCacheData(spDrawRichTextCache, rcDrawText, szScrollOffset, m_pTextData->GetTextRowXOffset(), GetAlpha());
                 m_pTextData->SetDrawRichTextCache(spDrawRichTextCache);
             }
         }
@@ -3120,7 +3120,7 @@ UiRect RichEdit::GetRichTextDrawRect() const
 
 uint8_t RichEdit::GetDrawAlpha() const
 {
-    return (uint8_t)GetAlpha();
+    return GetAlpha();
 }
 
 void RichEdit::OnTextRectsChanged()
